requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
fastapi==0.117.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
playsound==1.2.2
pydantic==2.11.9
pydantic_core==2.33.2
pypiwin32==223
pyserial==3.5
python-dotenv==1.1.1
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.2.3
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
